--- .vaadin.testbench.developer.license.txt
@@ -1,1 +1,0 @@
-f2c4ff51-f2e7-4d47-adb8-40100b20c125
