+++ src/main/java/org/amadeus/charon/ui/ReviewDisplay.java
@@ -1,4 +1,4 @@
-package org.amadeus.charon;
+package org.amadeus.charon.ui;
 
 import org.amadeus.charon.data.Review;
